server/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ds-catalog"
version = "0.0.0"
description = "Data Space Catalog"
readme = "README.md"
authors = ["HIRO-MicroDataCenters <all-hiro@hiro-microdatacenters.nl>"]
license = "MIT"
repository = "https://github.com/HIRO-MicroDataCenters-BV/ds-catalog"
homepage = "https://github.com/HIRO-MicroDataCenters-BV/ds-catalog"
packages = [{include = "*", from="app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109"
uvicorn = "^0.26"
classy-fastapi = "^0.6.1"
prometheus-fastapi-instrumentator = "^6.1.0"
polyfactory = "^2.16.2"
pydantic-settings = "^2.5.2"
neomodel = "^5.3.2"
pytz = "^2024.2"
types-pytz = "^2024.2.0.20240913"
aiohttp = "^3.10.5"

[tool.poetry.group.dev.dependencies]
black = "^23.12"
flake8 = "^7.0"
isort = {extras = ["colors"], version = "^5.13"}
mypy = "^1.8"
mypy-extensions = "^1.0"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.12"

[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.24.0"
pytest = "^8.2"
pytest-mock = "^3.12.0"
httpx = "^0.26"
tox = "^4.13.0"

[tool.isort]
# https://github.com/timothycrosley/isort/
line_length = 88

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.black]
# https://github.com/psf/black
line-length = 88
color = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
pretty = true
show_traceback = true
color_output = true
exclude = [
  '.git',
  '__pycache__',
  '.venv',
  'build',
  'dist'
]

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]
asyncio_default_fixture_loop_scope = "function"

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]
